Better convergence

+++ src/scalarfield.rs
@@ -6,7 +6,7 @@ pub struct ScalarField {
     pub nx: usize,
     pub ny: usize,
     pub nz: usize,
-    pub scalars: DVector<f32>,
+    pub scalars: DVector<f64>,
 }
 
 impl ScalarField {
@@ -19,7 +19,7 @@ impl ScalarField {
         }
     }
 
-    pub fn set(&mut self, x: isize, y: isize, z: isize, value: f32) {
+    pub fn set(&mut self, x: isize, y: isize, z: isize, value: f64) {
         if x < 0 || x as usize >= self.nx || y < 0 || y as usize >= self.ny || z < 0
             || z as usize >= self.nz
         {
@@ -34,16 +34,16 @@ impl ScalarField {
     }
 }
 
-impl Mul<f32> for ScalarField {
+impl Mul<f64> for ScalarField {
     type Output = ScalarField;
 
-    fn mul(mut self, rhs: f32) -> Self::Output {
+    fn mul(mut self, rhs: f64) -> Self::Output {
         self.scalars *= rhs;
         self
     }
 }
 
-impl Mul<ScalarField> for f32 {
+impl Mul<ScalarField> for f64 {
     type Output = ScalarField;
 
     fn mul(self, mut rhs: ScalarField) -> Self::Output {
@@ -63,9 +63,9 @@ impl<'a> Add<&'a ScalarField> for ScalarField {
 }
 
 impl Index<(usize, usize, usize)> for ScalarField {
-    type Output = f32;
+    type Output = f64;
 
-    fn index(&self, index: (usize, usize, usize)) -> &f32 {
+    fn index(&self, index: (usize, usize, usize)) -> &f64 {
         let (x, y, z) = index;
         assert!(x < self.nx && y < self.ny && z < self.nz);
         &self.scalars[x as usize + self.nx * (y as usize + self.ny * (z as usize))]
@@ -73,9 +73,9 @@ impl Index<(usize, usize, usize)> for ScalarField {
 }
 
 impl Index<(isize, isize, isize)> for ScalarField {
-    type Output = f32;
+    type Output = f64;
 
-    fn index(&self, index: (isize, isize, isize)) -> &f32 {
+    fn index(&self, index: (isize, isize, isize)) -> &f64 {
         let (x, y, z) = index;
         if x < 0 || x as usize >= self.nx
             || y < 0 || y as usize >= self.ny
@@ -89,18 +89,17 @@ impl Index<(isize, isize, isize)> for ScalarField {
 }
 
 impl Index<Point3<usize>> for ScalarField {
-    type Output = f32;
+    type Output = f64;
 
-    fn index(&self, index: Point3<usize>) -> &f32 {
+    fn index(&self, index: Point3<usize>) -> &f64 {
         &self[(index.x, index.y, index.z)]
     }
 }
 
 impl Index<Point3<isize>> for ScalarField {
-    type Output = f32;
+    type Output = f64;
 
-    fn index(&self, index: Point3<isize>) -> &f32 {
+    fn index(&self, index: Point3<isize>) -> &f64 {
         &self[(index.x, index.y, index.z)]
     }
 }
-
